requirements.txt
albucore==0.0.24
albumentations==2.0.8
annotated-types==0.7.0
certifi==2025.7.14
charset-normalizer==3.4.2
colorama==0.4.6
contourpy==1.3.2
cycler==0.12.1
defusedxml==0.7.1
distlib==0.4.0
filelock==3.18.0
filetype==1.2.0
fonttools==4.59.0
fsspec==2025.7.0
idna==3.7
Jinja2==3.1.6
kiwisolver==1.4.8
loguru==0.7.3
MarkupSafe==3.0.2
matplotlib==3.10.3
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.6
opencv-python==4.12.0.88
opencv-python-headless==4.10.0.84
packaging==25.0
pandas==2.3.1
pandas-stubs==2.3.0.250703
pillow==11.3.0
pillow-avif-plugin==1.5.2
pillow_heif==1.0.0
pipenv==2025.0.4
platformdirs==4.3.8
psutil==7.0.0
pure_eval==0.2.3
py-cpuinfo==9.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
requests==2.32.4
requests-toolbelt==1.0.0
roboflow==1.2.2
scipy==1.15.3
simsimd==6.5.0
six==1.17.0
stringzilla==3.12.5
supervision==0.26.1
sympy==1.13.1
torch==2.5.1+cu121
torchaudio==2.5.1+cu121
torchvision==0.20.1+cu121
tqdm==4.67.1
typing-inspection==0.4.1
typing_extensions==4.14.1
tzdata==2025.2
ultralytics==8.3.169
ultralytics-thop==2.0.14
urllib3==2.5.0
virtualenv==20.32.0
wcwidth==0.2.13
win32_setctime==1.2.0
